Move the re-enter password screen to screen folder root
